Refactor: Start breaking out tables again
I am using a relational db to manage my data so it should be normalized and not stored serialized in a single table.

However this means writing my own To and FromJSON instances in order to hide the CvId fields, which is a faff.

See http://lpaste.net/350901

This commit breaks out the database types into their own module and implements a references table that works as the others will in the future. Next a series of boring commits to make it all work this way...